Issue #12098 - Fix CpTaxonomyHelperTest grouping test
