+++ profile/modules/cp/modules/cp_taxonomy/tests/src/ExistingSite/CpTaxonomySettingsTest.php
@@ -5,7 +5,7 @@ namespace Drupal\Tests\cp_taxonomy\ExistingSite;
 use Drupal\Core\Cache\Cache;
 
 /**
- * Class CheckingFieldsTest.
+ * Class CpTaxonomySettingsTest.
  *
  * @group cp
  * @group kernel
@@ -19,13 +19,6 @@ class CpTaxonomySettingsTest extends TestBase {
   protected $renderer;
 
   /**
-   * The admin user.
-   *
-   * @var \Drupal\Core\Session\AccountInterface
-   */
-  protected $groupAdmin;
-
-  /**
    * {@inheritdoc}
    */
   public function setUp() {
@@ -33,7 +26,6 @@ class CpTaxonomySettingsTest extends TestBase {
     $this->renderer = $this->container->get('renderer');
     $configFactory = $this->container->get('config.factory');
     $this->configTaxonomy = $configFactory->getEditable('cp_taxonomy.settings');
-    $this->groupAdmin = $this->createUser([], NULL, TRUE);
   }
 
   /**
@@ -123,61 +115,6 @@ class CpTaxonomySettingsTest extends TestBase {
     /** @var \Drupal\Core\Render\Markup $markup_array */
     $markup = $this->renderer->renderRoot($render);
     $this->assertNotContains($term->label(), $markup->__toString());
-  }
-
-  /**
-   * Test group admin settings form.
-   */
-  public function testCpSettingsTaxonomyFormSelectNone() {
-    $this->configTaxonomy->set('display_term_under_content_teaser_types', NULL);
-    $this->configTaxonomy->save(TRUE);
-
-    $this->drupalLogin($this->groupAdmin);
-    $this->visitViaVsite('cp/settings/taxonomy', $this->group);
-    $this->assertSession()->statusCodeEquals(200);
-    $page = $this->getCurrentPageContent();
-    // Assert checkboxes are checked.
-    $this->assertContains('name="display_term_under_content_teaser_types[bibcite_reference:*]" value="bibcite_reference:*" checked="checked" class="form-checkbox"', $page);
-    $this->assertContains('name="display_term_under_content_teaser_types[media:*]" value="media:*" checked="checked" class="form-checkbox"', $page);
-
-    $edit = [
-      'display_term_under_content_teaser_types[bibcite_reference:*]' => 0,
-      'display_term_under_content_teaser_types[media:*]' => 0,
-    ];
-    $this->drupalPostForm(NULL, $edit, 'Save configuration');
-    $this->assertSession()->statusCodeEquals(200);
-    $page = $this->getCurrentPageContent();
-    // Assert checkboxes are unchecked.
-    $this->assertContains('name="display_term_under_content_teaser_types[bibcite_reference:*]" value="bibcite_reference:*" class="form-checkbox"', $page);
-    $this->assertContains('name="display_term_under_content_teaser_types[media:*]" value="media:*" class="form-checkbox"', $page);
-  }
-
-  /**
-   * Test group admin settings form.
-   */
-  public function testCpSettingsTaxonomyForm() {
-    $this->configTaxonomy->set('display_term_under_content', '1');
-    $this->configTaxonomy->set('display_term_under_content_teaser_types', NULL);
-    $this->configTaxonomy->save(TRUE);
-
-    $this->drupalLogin($this->groupAdmin);
-    $this->visitViaVsite('cp/settings/taxonomy', $this->group);
-    $this->assertSession()->statusCodeEquals(200);
-    $page = $this->getCurrentPageContent();
-    // Assert checkboxes are checked.
-    $this->assertContains('name="display_term_under_content" value="1" checked="checked" class="form-checkbox"', $page);
-    $this->assertContains('name="display_term_under_content_teaser_types[bibcite_reference:*]" value="bibcite_reference:*" checked="checked" class="form-checkbox"', $page);
-
-    $edit = [
-      'display_term_under_content' => '0',
-      'display_term_under_content_teaser_types[bibcite_reference:*]' => '',
-    ];
-    $this->drupalPostForm(NULL, $edit, 'Save configuration');
-    $this->assertSession()->statusCodeEquals(200);
-    $page = $this->getCurrentPageContent();
-    // Assert checkboxes are unchecked.
-    $this->assertContains('name="display_term_under_content" value="1" class="form-checkbox"', $page);
-    $this->assertContains('name="display_term_under_content_teaser_types[bibcite_reference:*]" value="bibcite_reference:*" class="form-checkbox"', $page);
   }
 
 }
